renamed styles file to style.js
