+++ src/components/header/header.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Background } from './headerStyles'
+import { Background } from './style'
 
 export default function Header({ children, ...props }) {
   return <Background {...props}>{children}</Background>
