+++ src/components/header/index.js
@@ -2,9 +2,9 @@ import React from 'react'
 import Header from './header'
 import Toggle from '../toggle'
 import { LayoutWrapper } from '../../styles/shared'
-import { HeaderContent } from './headerStyles'
+import { HeaderContent } from './style'
 import { useWindowSize, useLayoutType } from '../../hooks'
-import Searchbar from '../searchbar'
+import Searchbar from '../searchbar/desktop'
 
 import { ReactComponent as Logo } from '../../assets/desktop/logo.svg'
 import desktop from '../../assets/desktop/bg-pattern-header.svg'
